+++ diffsims/tests/test_utils/test_probe_utils.py
@@ -1,3 +1,21 @@
+# -*- coding: utf-8 -*-
+# Copyright 2017-2020 The diffsims developers
+#
+# This file is part of diffsims.
+#
+# diffsims is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# diffsims is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with diffsims.  If not, see <http://www.gnu.org/licenses/>.
+
 '''
 Created on 6 Nov 2019
 
@@ -6,7 +24,7 @@ Created on 6 Nov 2019
 
 import pytest
 import numpy as np
-from diffsims.utils.atomic_diffraction_generator_support.probe_utils import (ProbeFunction, BesselProbe)
+from diffsims.probe.probe_functions import ProbeFunction, BesselProbe
 
 
 def _toMesh(x):
